crop-recommendation-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
xgboost==2.0.3
shap==0.44.0
matplotlib==3.8.2
seaborn==0.13.0
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0



